refactor:  단순 Styled 컴포넌트 위치 이동
